svg loading feedback and abortion control
Most of the calculation time is spend rendering/loading the SVG.
Unfortunately aborting SVG loading once it started is not possible.

Rewriting the URL while it is being loaded would result in undefined
behaviour, probably even locking the browser.

To prevent this from happening at all, we check if the SVG is loaded
before starting the turtle worker inside the testCode function.

- loading feedback in the info field.
- disabled run/abort button during loading
- prevent running while loading