Switching between turtles / pages
Adding a menu to switch between turtles.
(For now just laser turtle and crafty turtle)

Each turtle corresponds to a github page.
The basic idea is, that you can jump between
turtle interpreters / github pages without losing your code
and see how the code is rendererd in a different world.
(i.e. the world of laser vs. the world of embroidery)

Each turtle should be able to handle the same commands,
but the results may differ.